+++ wp-content/themes/codigo/resources/scripts/app.js
@@ -17,6 +17,7 @@ import tinyslider from "./modules/tinyslider";
 import collaborators from "./modules/collaborators";
 import accordion from "./modules/accordion";
 import jobs from "./modules/jobs";
+import loadEffect from "./modules/loadEffect";
 
 /**
  * Application entrypoint
@@ -46,6 +47,7 @@ const CDG = {
     collaborators();
     accordion();
 
+
     window.addEventListener("resize", function(){
       //consoleHello('window has resized');
       if(window.innerWidth < 768){
@@ -60,6 +62,7 @@ const CDG = {
   onloadFunctions: function() {
     //consoleHello('CDG is loaded');
     jobs();
+    //loadEffect();
   }
 };